Modify parser
for문으로 동작하던 parser를 while문으로 변경 중
