--- src/type.js
@@ -1,8 +1,0 @@
-const DIVIDER = ["{", "}", "[", "]", ":", ",", "+", "-"];
-
-const SEPERATOR = ["{", "}", "[", "]", ":"];
-const OPERATOR = ["+", "-", "*", "/", "**"];
-const BOOLEAN = ["true", "false"];
-
-module.exports = { DIVIDER, SEPERATOR, OPERATOR, BOOLEAN };
-
